requirements.txt
asgiref==3.7.2
black==23.12.1
click==8.1.7
Django==5.0
django-admin-interface==0.28.0
django-colorfield==0.11.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
flake8==6.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-slugify==8.0.1
pytz==2023.3.post1
sqlparse==0.4.4
text-unidecode==1.3
